+++ documents-service/src/main/java/ru/alefilas/service/impls/DocumentServiceImpl.java
@@ -1,18 +1,23 @@
 package ru.alefilas.service.impls;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.PageRequest;
+import org.springframework.data.domain.Sort;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ru.alefilas.dto.InputDocumentDto;
-import ru.alefilas.dto.InputDocumentVersionDto;
-import ru.alefilas.dto.OutputDocumentDto;
-import ru.alefilas.dto.OutputDocumentVersionDto;
+import ru.alefilas.dto.documents.InputDocumentDto;
+import ru.alefilas.dto.documents.InputDocumentVersionDto;
+import ru.alefilas.dto.documents.OutputDocumentDto;
+import ru.alefilas.dto.documents.OutputDocumentVersionDto;
 import ru.alefilas.model.document.Document;
 import ru.alefilas.model.document.DocumentType;
 import ru.alefilas.model.document.DocumentVersion;
 import ru.alefilas.model.moderation.ModerationStatus;
+import ru.alefilas.model.moderation.ModerationTicket;
 import ru.alefilas.repository.DocumentRepository;
 import ru.alefilas.repository.DocumentTypeRepository;
+import ru.alefilas.repository.ModerationRepository;
 import ru.alefilas.service.DocumentService;
 import ru.alefilas.service.exception.DocumentNotFoundException;
 import ru.alefilas.service.exception.DocumentTypeNotFoundException;
@@ -20,43 +25,42 @@ import ru.alefilas.service.mapper.DocumentMapper;
 
 import java.time.LocalDate;
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class DocumentServiceImpl implements DocumentService {
 
-    private final DocumentRepository documentRepository;
-
-    private final DocumentTypeRepository documentTypeRepository;
+    @Autowired
+    private ModerationRepository moderationRepository;
 
     @Autowired
-    public DocumentServiceImpl(DocumentRepository documentRepository, DocumentTypeRepository documentTypeRepository) {
-        this.documentRepository = documentRepository;
-        this.documentTypeRepository = documentTypeRepository;
-    }
+    private DocumentRepository documentRepository;
+
+    @Autowired
+    private DocumentTypeRepository documentTypeRepository;
 
 
     @Override
     @Transactional
-    public OutputDocumentDto save(InputDocumentDto document) {
+    public OutputDocumentDto save(InputDocumentDto dto) {
 
-        Document savedDocument = DocumentMapper.dtoToModel(document);
+        Document document = DocumentMapper.dtoToModel(dto);
 
-        if (savedDocument.getId() == null) {
-            savedDocument.setCreationDate(LocalDate.now());
-            savedDocument.setStatus(ModerationStatus.ON_MODERATION);
-        } else {
-            Document docFromDb = findDocumentById(savedDocument.getId());
-            savedDocument.setStatus(docFromDb.getStatus());
-            savedDocument.setCreationDate(docFromDb.getCreationDate());
+        if (document.getId() == null) {
+            document.setCreationDate(LocalDate.now());
         }
 
-        DocumentVersion version = savedDocument.getCurrentVersion();
+        DocumentVersion version = document.getCurrentVersion();
         if (version.getId() == null) {
-            version.setStatus(ModerationStatus.ON_MODERATION);
-            savedDocument.addVersion(version);
+            document.setStatus(ModerationStatus.ON_MODERATION);
+            document.addVersion(version);
         }
 
-        documentRepository.save(savedDocument);
+        Document savedDocument = documentRepository.save(document);
+
+        if  (document.getStatus() == ModerationStatus.ON_MODERATION) {
+            createModerationTicket(savedDocument);
+        }
 
         return DocumentMapper.modelToDto(savedDocument);
     }
@@ -69,13 +73,13 @@ public class DocumentServiceImpl implements DocumentService {
         document.setCurrentVersion(documentVersion);
 
         if (documentVersion.getId() == null) {
-            documentVersion.setStatus(ModerationStatus.ON_MODERATION);
+            document.setStatus(ModerationStatus.ON_MODERATION);
             document.addVersion(documentVersion);
         }
 
-        documentRepository.save(document);
+        DocumentVersion savedVersion = documentRepository.save(document).getCurrentVersion();
 
-        return DocumentMapper.modelToVersionDto(documentVersion);
+        return DocumentMapper.modelToVersionDto(savedVersion);
     }
 
     @Override
@@ -87,6 +91,9 @@ public class DocumentServiceImpl implements DocumentService {
     @Override
     @Transactional
     public void deleteById(Long id) {
+        if (!documentRepository.existsById(id)) {
+            throw new DocumentNotFoundException(id);
+        }
         documentRepository.deleteById(id);
     }
 
@@ -109,8 +116,49 @@ public class DocumentServiceImpl implements DocumentService {
         return (List<DocumentType>) documentTypeRepository.findAll();
     }
 
+    @Override
+    @Transactional
+    public DocumentType save(String type) {
+        DocumentType documentType = new DocumentType();
+        documentType.setType(type.toUpperCase());
+        return documentTypeRepository.save(documentType);
+    }
+
+    @Override
+    @Transactional
+    public List<OutputDocumentVersionDto> getDocumentVersionsById(Long id) {
+        return findDocumentById(id).getVersions()
+                .stream()
+                .map(DocumentMapper::modelToVersionDto)
+                .collect(Collectors.toList());
+    }
+
+    @Override
+    @Transactional
+    public Page<OutputDocumentDto> getAllDocuments(int page) {
+        return documentRepository.findAllByStatus(ModerationStatus.CONFIRMED, PageRequest.of(page, 10, Sort.by("documentPriority")))
+                .map(DocumentMapper::modelToDto);
+    }
+
+    @Override
+    @Transactional
+    public Page<OutputDocumentDto> getDocumentsByType(String type, int page) {
+        return documentRepository.findAllByTypeTypeAndStatus(type.toUpperCase(), ModerationStatus.CONFIRMED, PageRequest.of(page, 5))
+                .map(DocumentMapper::modelToDto);
+    }
+
     private Document findDocumentById(Long id) {
         return documentRepository.findById(id)
                 .orElseThrow(() -> new DocumentNotFoundException(id));
     }
+
+    private void createModerationTicket(Document document) {
+        if (moderationRepository.findFirstByDocument(document).isEmpty()) {
+            ModerationTicket ticket = new ModerationTicket();
+            ticket.setDocument(document);
+            moderationRepository.save(ticket);
+        }
+    }
+
+
 }
